// types/settings.ts
export interface StoreSettings {
  storeName: string;
  logoUrl: string;
  faviconUrl: string;
  supportEmail: string;
  supportPhone: string;
  address: string;
  defaultShippingFee: number;
  freeShippingThreshold: number;
  codEnabled: boolean;
  upiEnabled: boolean;
  cardsEnabled: boolean;
  netBankingEnabled: boolean;
  instagram: string;
  facebook: string;
  youtube: string;
}